requirements.txt
asgiref==3.4.1
click==8.0.1
cycler==0.10.0
Django==3.2.7
image==1.5.33
imageio==2.9.0
importlib-metadata==4.8.1
joblib==1.0.1
kiwisolver==1.3.2
langid==1.1.6
matplotlib==3.4.3
nameparser==1.0.6
networkx==2.6.3
nltk==3.6.2
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
pdf2image==1.16.0
Pillow==8.3.2
pycountry==20.7.3
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
regex==2021.8.28
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
tifffile==2021.8.30
tqdm==4.62.2
typing-extensions==3.10.0.2
zipp==3.5.0